backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
reportlab>=4.0
whitenoise>=6.6,<7.0
gunicorn>=21.2,<22.0
